Add roles to user in `PopulateCurrentUser`
We now reuse `getUserRolesFromToken` to add the decoded roles from the
OAuth2 token to the user variable that's passed around the service.

This will allow us to read the roles at various stage in the service and
restrict access based on the user's permissions.

This commit adds mising tests for the `populateCurrentUser` middleware
to test both existing functionality and the addition of the user's
roles.