Rename OASys confirmation url
Although we tend to name urls after the resource/field name,
`oasys-confirmed` feels like a slightly incongruous name for end users
to see in the URL before they've actually confirmed that OASys data is
up to date.

Although this could potentially be seen as an inconsistency, I think
improving readability of the URL is more important here. I can't see
this decision causing confusion for future developers so I think this
inconsistency is OK.