Refactor prisoner client to take system token
We've now had the token added to our client credentials so we won't need
to set this on individual users. Rather than sending through the user's
token to allow them to access the endpoint, as we did previously, we now
send through the user's username through the system client token for
auditing purposes on the HMPPS Auth side of things.

This refactor allows us to authenticate with the Prisoner Offender
Search with our system client credentials rather than the logged in
user's token, meaning we don't need to add the `VIEW_PRISONER_DATA` role
to all of our users individually. That role is now on our own client.

There's a lot changing in this single commit here as we call the
`PersonService` in a few places so a few tests needed updating. The main
functional change here is the update to the `PersonService` to first
fetch the system token from HMPPS Auth and pass that (rather than the
user's own token) to the `PrisonerClient` to authenticate with the
Prisoner Offender Search.