Add `updateHasCourseHistory` controller action
I've added this to the referrals controller for now. Since course
histories hang off people, the `CourseHistoryController` should probably
live in the `refer/people` controllers directory. Conversely,
`hasCourseHistory` is a referral field, so it doesn't quite make sense
to put it in a controller for people's course histories. We could follow
the `oasysConfirmation` and `reason` pattern by adding a dedicated
controller for this form page, but a `hasCourseHistory` controller feels
weird