Exclude files from coverage checks
Per [an Approved Premises PR][1], most of these are template and/or
config files, plus some indices. We should be able to safely ignore
these

Some of these we have added more tests for, and function coverage is
not far from 100%, but I've chosen to ignore them for now. We can always
revisit this if we have time later

The next commit will turn on coverage checks

[1]: https://github.com/ministryofjustice/hmpps-approved-premises-ui/pull/42