Add course participation factories
The `CourseParticipationFactory` and the
`CourseParticipationOutcomeFactory` are following a new pattern. The
reason for this is that - based on the needs of the UI - they need to
have one of a few different combinations of properties.

A participation can either have a `courseId` or an `otherCourseName`,
but not both, and an outcome can't have a `yearStarted` if it's
complete, or a `yearEnded` if incomplete, or either if the status is
`undefined`

The extraction of some of the object generation from the fishery-based
factories helps achieve this (I couldn't see anything native to
fishery that supports this kind of functionality)