Add PrisonAPI Caseload type and factory
This represents the shape of the Caseload we'll get back from the Prison
API which we'll use to determine which People in Prison our POM and
Programme Team users have access to see. Users should only be able to
view people in a prison with a Prison ID that is in their list of
Caseloads.