Switch endpoint for fetching user's caseloads
The previous endpoint (`/api/staff/:staffId/caseloads`) wouldn't return
any caseloads for admin users (e.g. my own). It also meant we had to do
a slightly roundabout journey reading the logged-in-user's staffID,
whereas we can simply pass the user's token to the (apparently case-sensitive)
`/api/users/me/caseLoads` endpoint. This seems to allow us to fetch
caseloads even if the user is an admin.

We'll be using this to fetch the user's caseloads to pass them to the request to
the Prisoner Offender Search for a prisoner to ensure we're only
searching by the logged in user's caseloads.