Add `defaultMiddlewareOptions` to `RouteUtils.actions`
This allows you to pass in default `middlewareOptions` to the router,
which it will pass down as the default argument to the `get`, `post`,
and `put` actions, while still being overwriteable if you pass in that
argument. This reduces some of the clutter in route files, making them
a little DRYer and easier to scan

This doesn't change the approach to making routes open to all users -
the Find routes remain unchanged, for instance. The second argument to
`RouteUtils.actions` is optional will default to `undefined`, which
produces the same behaviour as before. This simply allows you to add
middleware options higher up the chain (at `.actions`) so that you
don't need to specify them on every `get`/`post`/`put` call