Only fetch users from HMPPS Auth on new session
Previously we were calling the `GET /auth/api/user/me` endpoint in HMPPS
Auth as part of every request when calling the `populateCurrentUser`
middleware.  We don't need to do this, as we can do it once the user
logs in, instead, by storing the user's details in the session and
checking the presence of this variable before making a call to `getUser`
if it's not set.

As part of this, now that we're adding `UserDetails` to the session
object, convert it to a type, rather than an interface, as importing it
as an interface to express types was causing a few strange issues when
typechecking. I suppose it makes sense for this to be its own exportable
type rather than an interface now.

I'm making this change now, as I'd like to set the user's caseloads here
too in future, which will require a call to the `Prison API` and I'd
like to avoid making that call on every single request.

This approach is adapted from the Temporary Accommodation team's
[here](https://github.com/ministryofjustice/hmpps-temporary-accommodation-ui/commit/70281bf82d01af5ce7ee9eafa9b13bb5458228cf)