Add caseloads to logged in user variable
Now that we're only making this request once per session, we can store
the user's caseloads in the user variable so we don't have to make an
additional request to the Prison API.

We'll be using these caseloads when searching for a Person in Prison via
the Prisoner Offender Search, on every page with the banner showing
personal details, so it makes sense to not have to make an additional
request every time.

If for any reason the logged in user doesn't have any caseloads, or an
error is thrown by when looking them up, we set the caseloads to an
empty array to prevent the user losing access to parts of the site that
don't require a caseload, e.g. Find.

The `userService` tests are getting a bit messy here, so I'll tidy them
up in the next commit.