Fix possible bug in `populateCurrentUser`
We previously overrode any `req.session.user` values with the older
values from `res.locals.user`, which felt a bit strange. The values
returned from `getUser` should be more up-to-date in theory (although
in practice there's probably not going to be much difference as they
both come from the same API).

I've added tests for this to make it a bit obvious what's supposed to
happen.

Updating this code highlighted a slight lack in understanding of how
this function worked. We believe that the `res.locals.user` check
at the beginning of the function is to allow this middleware to run in
isolation, as it would otherwise rely on values set in
`setUpAuthentication`, a separate middleware.