Display unanswered CourseParticipation fields
on the Programme History index page as "Not known".

Previously, we wouldn't show a row against a field that a user hadn't
filled in when creating a CourseParticipation. We now would like to
highlight this as "not known" to prompt users to enter any information
they do know, rather than ignoring the field. We don't want to make the
fields required, as this would prevent people who genuinely don't have
this information from being able to add anything.