Fix expectation
This was passing, but a later commit exposed this typo. As I understand it,
this should change the key within the object we're making an assertion about
having passed to a method. I have no idea how it passes without this
change, but it does break when combined with other changes to this action