Update docker-compose.yml
Add a second Postgresql database
Now have prison-register-db for prison-register-api and
accredited-programmes-db for hmpps-accredited-programmes-api.
Both databases listen on port 5432, but have different host names. The services are able to connect to the correct database server using their hostnames.
Database ports are also exposed to the host environment (your laptop) at 5432 for accredited-programmes-api and 5433 for prison-register-api

[Ynda edit: undid some of the formatting changes to keep it compliant
with Prettier]