Add `HmppsManageUsersClient`
Ideally this would have been broken up into smaller commits but it would have meant having failing tests on some of them.

The `userService` now refers to the new HMPPS Manage Users API to get a users `username` and then another call to get further details for that user, from the `username`.

It made sense to make a change to the user related types to match what is coming back from the manage users API, which then had a knock on effect to some other files.

It also meant making changes to the integration test `mockApis` in this PR too.