Add AuditService
This will be used to add messages to a configured AWS SQS queue.

For now, we do not require auditing to be configured, and we switch off
logging. We will change this once the SQS Queue is set up by the Cloud
Team, and the necessary environment variables have been set.

This approach is taken from the CAS team
[here](https://github.com/ministryofjustice/hmpps-temporary-accommodation-ui/pull/101/commits/42717adcc2981ce6c9cb8bb74169a8507b61ae1b)

The way we're exporting interfaces and classes in the `server/config.ts`
file differs from the normal grouping either using an index file or
grouping exports in a file, but because we were exporting classes and
interfaces, they were grouped separately at the bottom of the file which
didn't feel right. I'm happy for the individual exports to be an
exception from the norm in this file.