Add AuditMiddleware for logging audit events
This wraps around a request handler, sending an audit message using our
AuditService in addition to invoking the original request handler's
functionality.

We ensure there's a username sent in this audit message, logging an
error and redirecting to `/authError` if one isn't sent, a bit like we
do in `AuthorisationMiddleware` if the user doesn't have the correct
role.

Again, the approach here is taken from the CAS team's implementation
[here](https://github.com/ministryofjustice/hmpps-approved-premises-ui/pull/886/commits/139142b83c0605b420ce7fdae6cebbda2eaf6753)
and
[here](https://github.com/ministryofjustice/hmpps-approved-premises-ui/pull/886/commits/ab56b89e629484340062c4398c2efb3f870c5d4b).