Send audit messages based on redirect destination
By reading the "Location" header from the response, we can determine
where the request handler redirected the user to, and send audit
messages depending on this destination. Using the path pattern, we can
also determine the details to include in the audit message.

In many of our PUT and POST routes, we follow the pattern of
redirecting back to the source page in the event of an error, and to a
page for viewing the edited/created entity on success. We can use this
to send the appropriate audit event in each case.

This is fairly complex code, and I'd like to simplify it a bit if we get
a bit more time, I'm essentially treating it like a package we've taken
from the CAS teams
[here](https://github.com/ministryofjustice/hmpps-approved-premises-ui/pull/886/commits/60acca9dfa0c3427ec6b03cd3929e8385ac948e8),
as it's doing the job for them