Add `presentCourseParticipation` to `CourseService`
At the moment this feels like the best place for this kind of functionality. We had a brief discussion and decided that we're ok with services calling other services.

I did do a version where `getParticipation` made the call to `presentCourseParticipation` rather than in the controllers, but it became untidy pretty quickly with unrelated changes needing to be made to some controller methods that dont need to use the `CourseParticationPresenter` type. (e.g. `show` and `update` in `CourseParticipationDetailsController`).