Add auditMiddleware to routes
Here we pass in services to the routes middleware, so we can access the
auditSevice in multiple routes.

Next, we'll be specifying an audit event per-route.

Something's not quite right here with the lib check in our integration
tests, and I'm getting the following error:

```
node_modules/@redis/client/dist/lib/client/commands-queue.d.ts:27:5 -
error TS18028: Private identifiers are only available when targeting
ECMAScript 2015 and higher.

27     #private;
```

I've disabled it as it's incredibly hard to pin down what is
causing the issue. I believe it's something to do with importing types
from outside a `.d.ts` file when running the "es5" compiler option, but
I've been unable to fix this after spending some time investigating it.

However, we're already checking this in our main `tsconfig.json` so I don't see
there being a need to check it again in our integration tests.