Add CourseParticipation for deletion in test
Our setup on the API side runs tests in a random order, and doesn't
currently set up state for each test (a piece of technical debt we'd
like to address in future). Because of this, if the test for destroying
a CP is run before the test for fetching that CP, it will fail as the CP
will already have been deleted.