Add success/failure message at end of `script/test`
It's not easy to differentiate successful and unsuccessful `script/test`
runs at a glance, especially when it gets to the integration tests. Though
the presence of "All specs passed" for the Refer-disabled integration tests
should mean everything has run successfully, a lot of the different test
suite outcomes look pretty similar (on half a medium-sized screen) given
the spaced out black text and tables. If run in an editor, the real estate
would be even smaller. You might need to scroll up and/or look closely to
determine if all tests ran successfully

This adds a success or failure message (as appropriate) in a similar style to
`script/test-history` for easy confirmation