Add `stubReferralWithScenario` and `stubSubmitReferral`
We needed to stub multiple requests to get a referral when submitting a
referral to mock the different states a referral could be in.

Wiremock scenarios seems like the best option: https://wiremock.org/docs/stateful-behaviour/

[Edit by Ynda] By default, the scenario state will be `"Started"`, so the
first stub can use this as the `requiredScenarioState`. Providing
`newScenarioState` updates the state when the stubbed endpoint is hit, which
then allows stubs with different `requiredScenarioState`s to be matched on
future calls