Extract logic for presenting course participations
We're about to add the page for viewing course participations on
submitted referrals, so this will help simplify the code when adding
this functionality to a new controller, both in the implementation and
tests.

This is the first part of the refactor, we'll convert these sorted
participations into their summary lists in the next commit.