Update year in `CourseParticipationOutcomeFactory`
A helper method was included in this factory to generate a random
non-future year. We've since added validations to the year fields, but
not updated this method. We were passing in a minimum year of 1980, but
the minimum valid year is now 1990. This hardcodes that in the method

This was producing flaky results when asserting against success messages
in integration tests, since the factory would occasionally produce
invalid course participation outcomes that would block success
conditions