Update Refer nesting
We previously moved to nest the old `/referrals` paths as
`/refer/new/referrals`, so that the submitted referrals could use the
barer path `/refer/referrals`. This updates the rest of the codebase to
follow this nesting, with submitted referrals the default/top layer (or
occasionally nested under `show` where appropriate)

The paths file is slightly changed, so that we aren't created `const`s
where nothing inherits from it and its value isn't used multiple times.
This reduces the text walls at the top a little