Create Participation summaryList in service method
This extracts the logic for converting Participations into summary list
options into the `presentCourseParticipation` function in the Service,
rather than relying on multiple functions being called in the
controller.

We might want to rename this method to be more specific about the fact
that it's converting to a summary list, but I suppose we're still
"presenting" it.