Fix empty sentence start date displaying as today
We refactored the `DateUtils.govukFormattedFullDateString` function to
either format the date passed in, or return today's date. This meant
that we were incorrectly showing today's date on the sentence start date
page field here.

We'll hopefully have some nicer content to show here than an empty row,
but this at least means we're not displaying inaccurate information.