Use Prisoner Search for sentence start date
We previously identified Prison API endpoints to get the sentence start
date and various release dates. However, the Prisoner Search already
calls the Prison API and includes the dates we need in its prisoner
model. We will therefore soon be using the prisoner model to get the
release dates (rather than adding an additional Prison API endpoint and
making an extra request), so we've decided here to get the sentence
start date from the prisoner model too (rather than sentences and
offences) to keep all dates together. We still need the Prison API
sentences and offences call to get the sentence type description, which
isn't included in the prisoner model on the Prisoner Search

Most of the prisoner model data is optional in the API model. We should
update our model properly at some point, but for now, this new field is
added as optional while older fields are left with the assumption that
they will be present

The relevant util method is updated to reflect the new sentence date
source and deal with the possibility of it being `undefined`

Co-authored-by: George <george@dxw.com>