Add new SystemToken type
There are two types of tokens being passed around in our service, which
may be misleading. We were (incorrectly) referring to the User token
type in the PersonService, when we're actually passing the HMPPS Auth
System token, which has all of the client's roles on it. This should
hopefully clarify which tokens are being used where, even though they're
basically just a wrapper around a `string`.