Add `releaseDatesSummaryListRows` utils method
This is pretty complex and might need breaking down a little. Missing
dates are currently being tested in the condition of all dates missing.
Eventually this state will never be rendered - a separate piece of work
will handle showing something else in that condition - but this util
doesn't necessarily need to know about that