Add `findOffencesThatStartWith` to `PrisonApiClient`
Have mirrored the naming from Prison API (https://prison-api-dev.prison.service.justice.gov.uk/swagger-ui/index.html#/offences/getOffencesThatStartWith) but replaced `get` with `find` as we do in the clients.

To get further detail for an offence, this is seemingly the best endpoint to get further information for an offence from the offence code. Unfortunately this endpoint returns a paginated result set with an array of results in `content`.  We will be using the full `offenceCode` so the number of results should be limited to a single item, so when use this in the upcoming service(s), we can select the first item from the `content` array.