Add `indexRedirect` to `CaseListController`
When we link to to the case list page, the user needs to be shown the correct case list which should be the first course in the list of courses for an organisation, sorted in alphabetical order.

So when a user visits `/assess/referrals/case-list` they're taken to `/assess/referrals/a-course-name/case-list` or shown the not found page if there are no courses for that organisation.