fix: trello(1198) handle missing user loading
When a user is not found a message is returned. The behaviour for two controllers was to throw if the user were missing.

Note the original function and its behaviour still exists as it may be used by other work in progress, and that behaviour may be preferred. If this is not the case a separate task could be actioned to remove it.