(1198) Add Basic Refer API Stubs
As part of 1198 I attempted to mock the behaviour of a missing referer on local dev. Half way through a link came up on cloud dev which could be used to verify the ticket change and I also realised the refer dashboards are in a state of flux. This stub work is enough to get the refer overview up and running as it currently stands, with the four states (one for one org, three for another).