Stub referral summaries
This adds referral summaries and courses by organisation to the
Accredited Programmes API stubs

When we were working with data we own served by the internal API, we
could seed it via the API. When we were working with data we don't own
served by external APIs, we could stub it in the UI using Wiremock.
We're now in a situation where the internal API serves both data we own
and data we don't, which makes generating fake data for local
development a bit trickier. For now, this stubs the internal API
endpoint to get around that - it means we lose the benefits of the
generated referral seeds when mocking locally (at least short term), but
we can smoke test features that require referral summaries with both
internal and external data

The referral summaries stubs are using string interpolation on the URL
rather than Wiremock's query parameters functionality. Without this -
and also URL encoding the spaces and commas - Wiremock would report a
mismatch between the stubbed and requested URLs. For some reason, this
doesn't appear to be a problem in the integration tests