Add additional fields to `ReferralSummary`
Most of these are from updates to the API to expose external data.
`tasksCompleted` was added recently in the `ReferCaseListController`,
and will be needed in the type so that it can be used when generating
table rows

`Object.prototype.hasOwnProperty.call` is being used here to retain
`undefined` values that are passed into the factory build method.
Otherwise `undefined` values would be ignored and the (e.g.) faker
methods would be used to generate random data instead