Allow API to provide response for external data
The Accredited Programmes API now makes requests to the Prisoner Search
and Prison Register. As we can't run the Prisoner Search locally, we
need to provide some wiremocked responses.

I've spent a lot of time trying to get our UI to use the wiremock
mappings in the API codebase, but I think this will require some more
detailed changes to the Dockerfile. This at least allows us to get
something up and running locally.

We're using the local docker container for the Prison Register here, as
that's already running in our UI codebase.