Use Refer paths in Refer case lists
Previously we were using the Assess paths everywhere for the link to
view an individual referral. This currently assumes `assessPaths` as the
default without a particularly valid reason, but providing a default
allows us only to need to pass the paths in to `tableRowContent` when
asking for that specific field (rather than every field). A refactor
could break down `tableRowContent` into multiple methods, but that would
result in a utils file bloated with small table row content methods