Add Refer case list integration tests
We could theoretically have used `sharedTests` for these so that Assess
and Refer tests are more in line, but I felt that the level of overlap
between the two test files didn't really warrant that. The main shared
logic is the use of an abstract method to check the table. The setup and
the non-table assertions are different enough between the two that it's
probably best not to take the DRYing up/abstraction any further here

One thing that could be DRYed up is the pagination testing, but for now
it'll probably do the job