Make optional fields optional in factory
The `OffenceDetailFactory` currently doesn't reflect the optionality of
the `OffenceDetail` fields. This doesn't appear to require any test
changes at present, but if we add in `OasysClient` tests later we'll
likely need to add a `withAllOptionalFields` trait to avoid
unpredictable nullish data

This also adds a `withAllOptionalFields` trait, similar to other
factories, for when we want to build an object with no nullish optional
fields. This is used in a controller test to ensure that `when`-based
mocking for `OffenceAnalysisUtils.textValue` still works. We use `when`
to mock the return value of a method when passed particular arguments:
if two fields are both `null`, the second `when` `null` mock would
overwrite the first, which creates unpredictability