Migrate from audiences to audience
We're moving to a single `audience` string on a course instead of an
array of strings. This updates the types and then the rest of the
codebase to match

The audience typing is tightened in `CourseAudience` itself, but
loosened back to `string` in query parameter contexts since those aren't
necessarily in the same format. The same is done for the course query
parameter, which was only typed as a `string`, but logically the same
applies

The `generateApiSeeds` script removes the `audience` and
`course_audience` tables. Technically these are still needed, but they
will be removed after this UI work is deployed. For now, they've been
manually added back into the generated seeds on the API side