Update referral client, service and api path to new dashboard endpoints
Use new endpoint, will raise a follow on PR to remove use of the now deprecated endpoint.

[Squash] Update CaseListControllers to use API sorting

We no longer want to sort on the client so this will take the browser URL request query parameters and pass those onto the API to get the correctly sorted results.

[Squash] Update case list integration tests to use `ReferralView`

This does not include any new tests, will add those in a future PR.