Update withdraw pages to use `previousPath` value when neccessary
The withdraw journey can be started from the the status history page and the update status journey, so the withdraw category page needs to know where to go Back to.

The reason information page also needs to have a flexibile back button as it can be redirected to from both the category and reason steps, depending on your selection.