Move initial update status step to be for assess journey only
The select decision step of the update status journey is only for users on the assess journey, so this has been moved from the shared directory to assess.  Paths have also been updated so they no longer handle potentially being on the refer path.

Some new paths for the next step have been added in anticpation of the addition of the next step where a user confirms their decision with eitheir a confirmation checkbox or reason text entry step.