Delete `referralStatusUpdateData` in session when viewing status history
This is generally the entry point for updating a status and the page that gets redirected back to if there is an issue during the journey.  Resetting it when visiting this page reduces the possibility of the session getting into a state where the data is for another referral, or has the incorrect values.