Add `CategoryController`
This will replace the current `WithdrawCategoryController` and allows us to share the code with other statuses that need to select a category when updating a status, such as deselected.

Content is in the controller for the moment but there is a discussion to move this to the API, so new statuses can be added without a further change being required in this app.