Use course id instead of names in the case list view and navigation
This worked when the course names had no special characters in and each word in the name was capitalised.  Encoding was an option but then that left us with the issue of still not knowing the correct casing of the course name.

Much safer to just use the course ID and update the url so they're prefixed with `/course`, so not to be confused with a referral id.