Handle the deselect and keep open status update journey
This is an exception to all the other possible status transitions, in that there are 2 `DESELECTED` statuses returned from `/status-transitions/`.  One of which is slightly different as it will have a `deselectAndKeepOpen` property. When it does, we alter the `code` value slightly and append `|OPEN` so the radio button has a different value.  We can then use this value to maintain that selection (`initialStatusDecision`), when navigating between pages, but also later down the line to redirect BACK to the decision step from the select a reason step, but this time it will show some different status transitions that the referral should actually be moved too (`finalStatusDecision`), before progressing to the final selection step.