Update mock api stubs
These had been a bit neglected lately so a few of the pages didn't work when running locally against the mock api.  This should fix things up a bit.

The status update journeys still won't though and I suspect it would be a bit of work to get them working to mimic the actual API.