Add `CASE_TRANSFER_ENABLED` env for feature flag
For a short period, we want to allow users with both refer and PT roles to be able to search for prisoners that are not in any their caseloads. We decided that a feature flag would be a good option to allow us to easily enable and disable this functionality, as and when.