Add `RecommendedPathwayController`
This requires `req.session.pniFindAndReferData?.prisonNumber` from the previous person search step.  It will then retrieve the persons details along with their PNI information and present it in a similar way to the existing PNI page within a referral. It is worth mentioning that the content and it's appearance is slightly different between the existing and this new page.