Use `pniFindAndReferData.prisonNumber` if available when raising a referral
When raising a referral using the new pni find journey then we need to carry the previously entered prison number through so the user doesn't get asked to enter it again.

We also need to reset the `req.session.pniFindAndReferData` object when viewing the current find directory so the user still has the option to search using the existing journey.