Updates for switching to the PNI Find and Refer journey
Please note: I appreciate this is a less than ideal commit size, but I’m not sure there is a nicer way break it down without having failing tests on each commit.

This commit is required as a result of moving solely to the PNI Find and Refer journey. This has meant that some pages and content (e.g Make a referral button) should no longer be accessible unless going through the process of the PNI Find and Refer journey. Also, because some pages now have multiple entry points, we needed to make sure any back or cancel links on those pages were correct within the context of what the user is doing.

- Move conditional login for showing Make a referral button from template into controller. Now unit tested as well as integration tested.
- Remove `NewReferralsPeopleController.find method`.  This  has been replaced `PersonSearchController.submit`
    - Also updated the `NewReferralsPeopleController.find.show` method to redirect back to `PersonSearchController.show` on error
- Removed `NewReferralsController.new` method as this has also been replaced by the `PersonSearchController.show`
- Updated `NewReferralsController.start` method to pass hrefs to the template, rather than create them in the template
    - Also updated to return user back to `PersonSearchController.show` if there is no prison number in `session.pniFindAndReferData`
- Added allowedRoles to the middlewareOptions within the find routes to restrict access to those routes
- Add integrations tests to for PNI Find and Refer to cover submitting a referral and encountering a duplicate
    - Any integration tests that have been removed should now be covered within here.