Add `courseVariantId` to BuildingChoices session data
We need to use the appropiate building choices course id for navigating back to the correct page.  Without this, the user would end up back on the generic course listing page and not the building choices specific one, so they would see the incorrect offerings/locations and unrelated to the options selected on the building choices form.

Also repopulate building choices form values when navigating back.