Switch to use PeopleSearchResponse type from our api
We have been referring to the Prisoner type from the Prisoner search api for a while now, despite using a similar but different type from our own API.

It was on the radar but we didn't have an immediate need to update it. However since the addition of the required prisonId property from the PeopleSearchResponse, it's time to switch to it.