Add `TransferReferralController` with `show` method
This will check if the referral can be transferred (currently doesn't include checking for duplicates). If it can be transferred then it will show the reason form.

If it cannot be transferred then it will set some error data in the session and redirect to an upcoming Transfer Error page.