APG-712: Formalise the encoding/decoding of audience and hasLdc params
**WHAT**

This commit introduces centralised code on CourseUtils which combines
hasLdc and audienceName into a single string, separated by ::

**WHY**

As per design's request, the case list page should show options for
Building Choices courses with "all cases" and "LDC only" cases.

To do this, we need some way of combining these values and encoding them
into a string, in order to:

1. Represent this in the UID
2. convert them back to a string and a boolean (possibly undefined)

This allows the code from the API Client -> Controller to consider the
values as a string and boolean pair, but from Controller -> UI to
consider them as a single string.

This is ultimately a pragmatic choice, it may well be that we decide
that storing these parts of state in a single string is a bad idea - and
we welcome any future development work which may improve it.

**REFACTORING**

Moved the case conversion (toLowerCase) for checking if an option item
is selected from the caller of the function, into the body of the
function itself.

This makes it impossible to forget to call `toLowerCase` and makes our
code a little bit more friendly.