Remove `organisationEnabled` from `CourseOffering` type
Would have preferred to use the generated `CourseOffering` type from the swagger API. However `id` can be marked as `null` as the same type is used for the creation of a course offering.

Have discussed and suggested that we add a type type along the lines of CreateCourseOffering which omits an ID and then make CourseOffering.id non nullable.